% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab1_quickBox.R
\name{Lab1_quickBox}
\alias{Lab1_quickBox}
\title{Create a boxplot}
\usage{
Lab1_quickBox(values)
}
\arguments{
\item{values}{Expects a vector of values.}
}
\value{
Boxplot
}
\description{
Create a boxplot
}
