% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab2_zScore.R
\name{Lab2_zScore}
\alias{Lab2_zScore}
\title{Performs finds z score}
\usage{
Lab2_zScore(values)
}
\arguments{
\item{values}{Expects a vector of values}
}
\value{
z-score vector
}
\description{
Performs finds z score
}
