% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab3_quickPlot.R
\name{Lab3_quickPlot}
\alias{Lab3_quickPlot}
\title{Creates a quick plot with steel blue dots}
\usage{
Lab3_quickPlot(xVar, yVar, title, xlabel, ylabel)
}
\arguments{
\item{xVar}{x-axis variables}

\item{yVar}{y-axis variables}

\item{title}{main title}

\item{xlabel}{x label}

\item{ylabel}{y label}
}
\value{
returns plot
}
\description{
Creates a quick plot with steel blue dots
}
