% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab4_fitVsResPlot.R
\name{Lab4_fitVsResPlot}
\alias{Lab4_fitVsResPlot}
\title{Creates a plot of residual vs fitted}
\usage{
Lab4_fitVsResPlot(dat, x, y)
}
\arguments{
\item{dat}{data table}

\item{x}{x variable}

\item{y}{y variable}
}
\value{
returns plot
}
\description{
Creates a plot of residual vs fitted
}
