% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab5_mybin.R
\name{Lab5_mybin}
\alias{Lab5_mybin}
\title{Creates a barplot which displays the relative frequency
of randomly selecting a number from 1-10.}
\usage{
Lab5_mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{number of iterations}

\item{n}{number of trials}

\item{p}{probability of success}
}
\value{
returns barplot and the values in each bin
}
\description{
Creates a barplot which displays the relative frequency
of randomly selecting a number from 1-10.
}
