% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab6_myncurve.R
\name{Lab6_myncurve}
\alias{Lab6_myncurve}
\title{Creates a density plot using dnorm given a mean and standard deviation.w
It finds the probabilty from -inf to a}
\usage{
Lab6_myncurve(mu = 10, sigma = 5, a = 6)
}
\arguments{
\item{mu}{mean}

\item{sigma}{standard deviation}

\item{a}{-inf to a}
}
\value{
returns density plot with the area colored and prints out probability
}
\description{
Creates a density plot using dnorm given a mean and standard deviation.w
It finds the probabilty from -inf to a
}
