% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab8_mycltb.R
\name{Lab8_mycltb}
\alias{Lab8_mycltb}
\title{CLT Binomial}
\usage{
Lab8_mycltb(n, iter, p = 0.5, ...)
}
\arguments{
\item{n}{number of elements}

\item{iter}{number of iterations}

\item{p}{probability of success}

\item{...}{additional parameters}
}
\value{
returns density plot with a theoritical normal curve and a combined vector of sums
}
\description{
CLT Binomial
}
