% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Creates a bootstrap density plot with confidence intervals}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of resamples}

\item{x}{sample data}

\item{fun}{statistic being used}

\item{alpha}{confidence interval}

\item{cx}{graph label scale}

\item{...}{extra parameters}
}
\value{
returns density plot with confidence interval
}
\description{
Creates a bootstrap density plot with confidence intervals
}
