% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{Findings maximum likelihood}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{statistic being used}

\item{x}{data}

\item{param}{parameter values}

\item{...}{extra parameters}
}
\value{
returns plot showing a log likelihood curve
}
\description{
Findings maximum likelihood
}
