% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project1_ntickets.R
\name{ntickets}
\alias{ntickets}
\title{ntickets
calculates the number of tickets to be sold when the number of seats in the flight}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{number of seats}

\item{gamma}{probability of overbooked}

\item{p}{probability of a show}
}
\value{
returns discrete and continuous plot
}
\description{
ntickets
calculates the number of tickets to be sold when the number of seats in the flight
}
